Dependencies Housekeeping (#129)
* Use AWS-specific fog gem instead of metagem

The `fog` library provides a metagem that installs gems for *every*
single provider it supports. People Finder only uses AWS, so there is
no reason to have any of the other gems included.

* Update `delayed_job` gem

This gem is pinned to a 4 yeard old Git commit - it should be a proper
gem version, and the most up to date one at that. The syntax for options
has changed since and needs updating too.

* Update `carrierwave` gem

This was pinned at an ancient Git revision, and none of the released
versions around that commit seemed to work. This updates `carrierwave`
to 1.0.0, which still needs further updating later but works.

- Update file type validation to new format in 1.0.0
- Add translation that is no longer provided by default

* Update various gems to fix deprecation warnings

* Update `simplecov` gem

The new version of this code coverage tool seems to measure coverage
differently, leading to a sudden drop in coverage from ~95% to ~70%.

- Decrease required coverage level for now until coverage can be
  improved

* Update `brakeman` gem

The new version finds some extra warnings, but these are all benign.

* Update `factory_bot` gem

The `factory_girl` gem dependency has been renamed and needs to be
changed to new name.