Mark flaky search specs as pending (#132)
These specs consistently fail when the test suite is run in a certain
order (via --seed).

Attempted to try and figure out what is going on, but there seems to be
a whole lot of things going wrong with the search specs (as evidenced by
the numerous specs already marked pending). Parking this until we have
time to properly triage and evaluate all the failures.