Dockerise app for development; new README
The previous Docker setup for the app was geared towards MoJ production
deployment. This adds a new `Dockerfile` and `docker-compose.yml`
suitable for local development, and makes a few changes to the app
configuration to allow running the app in this context.

It also moves the existing README to the `docs/` folder and adds a new
README with the information required to get the app up and running.