Update Ruby to 2.5.3 (#144)
* Upgrade Ruby to 2.5.3

* Remove `simplecov-rcov`

This gem is not compatible with Ruby 2.5, and hasn't been updated in
almost 10 years. There's no reason for us to use a specific SimpleCov
formatter (perhaps MoJ needed that for Jenkins?) so we can happily live
without it.

* Fix Poltergeist logging error on Ruby 2.5

Looks like `File::NULL` used to return a file object, and now returns a
`String`. We can work around this by just using a `StringIO` for the
null logging.

* Bump CF Ruby buildpack