Remove special handling of full name searches
People Finder handles the special case of searching for someone's full
name in a peculiar way - not only does it assume that if you are
searching for two words (no more, no less) that constitutes a full name,
it then goes and only displays *one* search result. Sounds great in
theory, but sometimes two people have the same name!

There needs to be some proper work at some point to clean up the
`PersonSearch` code and adjust the weightings, but for now this will
produce more meaningful search results.