Render descriptions in team lists without hyperlinks
Some team descriptions contain hyperlinks, which is an issue because
they are rendered into HTML as part of the links to teams on team list
pages, and the <a>-within-<a> breaks the page rendering in weird ways.

Govspeak won't let us specify a tag white/blacklist, so instead, this
just `gsub`s the offending <a> tags out of the Govspeak output (ugly
but does the job).