Clean up unused ex-MoJ functionality
People Finder was forked by DIT from an MoJ project. As part of bringing
it into DIT, a lot of functionality was either deactivated or
superseded, but the underlying code was never removed due to time
pressure and uncertainty about whether it would be used again in the
future.

This removes the vast majority of this code with the aim of making
development on People Finder slightly less painful.

Among other things, it removes:
- **Token-based authentication** - DIT uses a single sign on system to
  authenticate People Finder users, rendering the old way of signing in
  by having a link sent to you pointless). Includes removing:
  - Read-only users
  - Permitted domain logic
  - IP address whitelisting
- **Unused notification emails** - These have been deactivated but never
  fully removed
- **Problem Report mailer** - this now gets sent to Zendesk via the API
- **Information requests/suggestions** - this has been disabled at DIT
- **Organogram display** - this was half-baked and has been disabled at
  DIT - it will probably be reimplemented properly at some point