Use DIT SSO user_id to find user on login
This is the first step towards using DIT Staff SSO's `user_id` field
instead of the user's email as the primary way of identifying a user on
login.

- Before attempting to find a user by their SSO email on login, we check
  if a user exists with the SSO `user_id`. Only if that is unsuccessful
  do we try to find the user by email.
- On both login and user creation, store the SSO `user_id` in a new
  `ditsso_user_id` column against the People Finder user, so that all
  new users (and all those who log in after this changes is deployed)
  will have their SSO `user_id` associated with their profile.

The next step will be creating and running a task to populate all
existing users with `user_id`s from SSO. Once that is done, the
finding by email logic can be removed (and the rest of the login
code refactored).