Remove "internal auth key" based login
We should now exclusively use the Staff SSO User ID to identify users
when creating a session. This removes all of the `internal_auth_key`
logic, with the exception of the DB column (which will be removed in a
later commit once the changes have been without issues for a few days).

- Remove accounts Rake task
- Remove internal auth key from profile extracts CSV
- Remove internal auth key from factories
- Remove fetching by IAK in API; remove AuthUserLoader
- Remove internal auth key based login
- Refactor `SessionsController` to simplify the way it works