Use Ruby 2.6.5 and custom Dockerfile
The Docker image used to facilitate the change from PhantomJS to
Headless Chrome is a few minor Ruby versions behind and does not
seem actively maintained. It doesn't do anything extremely
exciting to begin with, and is just 10 odd lines of code on top
of the default Ruby image, so this brings it into this repo.

- Update Ruby to 2.6.5
- Use latest Headless Chrome and Chromedriver