Add initial basic Webpacker integration
We need to move the frontend off of ancient GOV.UK Elements on to the
modern GOV.UK Design System/Frontend. As part of that, we also want to
migrate away from the Rails asset pipeline towards Webpacker. It turns
out these two can coexist happily in a single application, so we are
adding Webpacker and will start slowly moving the frontend over one view
template at a time over the coming weeks.

- Add Webpacker gem and set up basic configuration
- Add initial templates and styles from Wagtail alpha
- Rename existing templates/assets to "legacy"
- Add working header and footer