Remove IP and browser info from user-facing audit trail (#290)
This doesn't really add value to users - we don't need to display the
browser a user made a change from or what their IP is. We log this stuff
anyway through other mechanisms, and we can keep it around in the DB for
now as `paper_trail` tracks it anyway, but we shouldn't show superfluous
information for the sake of it.