Don't fail login if user record fails to save
The move to the new `LoginPerson` interactor changed the method of saving
from `#save` to `#save!` because the operation shouldn't really fail.

However, in cases where there is an existing inconsistent record in the
DB, this _can_ fail and cause an error, so we need to catch it (and ship
it off to Sentry so we can investigate those accounts where that happens).