Hotfix for empty profile image API issue (#362)
The changes in the new photo upload flow seem to have broken the profile
image API which no longer returns a URL for default images (just an
image name). This forces the API to return `nil` when the profile image
URL doesn't seem to be a URL (doesn't start with 'http') so Digital
Workspace doesn't 500 for users without profile images.